revert part of 5cd163969b0958bd5812f837ff014409bf5b72ce
size_t is unsigned type, but function returns negative value
(function is apparenly not used anywhere else)

+++ l10ntools/inc/xmlparse.hxx
@@ -176,7 +176,7 @@ public:
     );
 
     virtual int GetPosition( ByteString id );
-    size_t RemoveChild( XMLElement *pRefElement );
+    int RemoveChild( XMLElement *pRefElement );
     void RemoveAndDeleteAllChilds();
 
     /// returns a child element which matches the given one
